Small collection of (3D) tool scripts
A few tools I've made in my time working with godot. These are all for 3D, but can easily be converted if we decide to work in 2D.
ColShapeScaleFixer: Fixes an issue with scaling on collision shapes, primarily on rigidbodies. (Effectively applies the scale to the shape as opposed to putting it in the transform matrix)
Freelookcamera: Pretty self explanatory. Ports the editor's freelook into the game
UniqueModelFixer (And children): Fix an issue with .obj imports not having unique materials, and having incorrect material properties.